requirements.txt
ajsonrpc==1.2.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
bottle==0.13.1
cachetools==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
docker==7.1.0
dotenv==0.9.9
et_xmlfile==2.0.0
executing==2.2.0
Faker==37.4.2
fastapi==0.116.1
fastjsonschema==2.21.1
fonttools==4.58.4
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.14.0
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
marshmallow==3.22.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mysql-connector-python==9.4.0
narwhals==2.0.1
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.5
pdfkit==1.0.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
platformio==6.1.16
plotly==6.2.0
prompt_toolkit==3.0.52
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyelftools==0.31
Pygments==2.19.2
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.0.2
referencing==0.36.2
reportlab==4.4.2
requests==2.32.3
rpds-py==0.26.0
seaborn==0.13.2
semantic-version==2.10.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.1
streamlit==1.47.1
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.30.6
watchdog==6.0.0
wcwidth==0.2.13
wsproto==1.2.0
